+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Update `cmd/hidden_lake/applications/filesharer`: fix creating hlf.stg
 - Update `cmd/hidden_lake/applications/filesharer|messenger`: fix tests with same ports
 - Update `cmd/hidden_lake/helpers/adapters/common`: fix print log
+- Update `cmd/hidden_lake/applications/filesharer`: fix dockerfile, storage path
 
 <!-- ... -->